+++ bridge/__init__.py
@@ -2,4 +2,6 @@
 
 __all__ = ["__version__"]
 
-__version__ = "0.1.0a3"
+__version__ = "0.1.0a4"
+
+